requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
tqdm>=4.50.0
torch>=1.9.0
transformers>=4.20.0
boto3>=1.26.0
pillow>=9.0.0
base64io>=1.0.0
streamlit>=1.22.0
watchdog>=2.1.0